Move FFI logic into its own file
